% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_projected_xy.R
\name{attach_projected_cols}
\alias{attach_projected_cols}
\title{Attach reprojected XY columns to tibble with coordinates for point data}
\usage{
attach_projected_cols(
  data,
  coords,
  from_crs,
  to_crs,
  to_coords = c("X", "Y"),
  remove = TRUE
)
}
\arguments{
\item{data}{data.frame / tibble with columns containing lat/lon or projected XY coordinates.}

\item{coords}{in case of point data: names or numbers of the numeric columns holding coordinates}

\item{from_crs}{coordinate reference system of the current coords columns, integer with EPSG code, or character with proj4string ... something suitable as input to \link[sf]{st_crs}}

\item{to_crs}{desired coordinate reference system for new coords columns,integer with EPSG code, or character with proj4string ... something suitable as input to \link[sf]{st_crs}}

\item{to_coords}{names for the new columns of reprojected coordinates; if \code{remove = FALSE}, these should be different from \code{coords}}

\item{remove}{logical; remove the original coords columns from the data?}
}
\value{
A tibble with all the non-spatial columns in \code{data} plus two new columns containing reprojected point coordinates.
}
\description{
Attach reprojected XY columns to tibble with coordinates for point data
}
\examples{

}
