% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWDutils-package.R
\docType{package}
\name{AWDutils-package}
\alias{AWDutils}
\alias{AWDutils-package}
\title{AWDutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Adam Davis \email{davisadamw@gmail.com} (\href{https://orcid.org/0000-0001-6324-6380}{ORCID})

}
\keyword{internal}
